+++ Frontend/src/lib/images/index.ts
@@ -1,15 +1,29 @@
-export {default as icon} from './icon.svg';
-export {default as logo} from './logo.svg';
-export {default as logoBeside} from './logo-beside.svg';
+export {default as icon} from './logo/icon.svg';
+export {default as logo} from './logo/logo-white-text.svg';
+export {default as logoBesideText} from './logo/logo-beside-text.svg';
 
-export {default as github} from './github.svg';
-export {default as simplex} from './simplex.svg';
-export {default as tor} from './tor.svg';
+export {default as github} from './links/github.svg';
+export {default as simplex} from './links/simplex.svg';
+export {default as tor} from './links/tor.svg';
 
-export {default as info} from './info.svg';
-export {default as alert} from './alert.svg';
-export {default as success} from './success.svg';
+export {default as info} from './status/info.svg';
+export {default as alert} from './status/alert.svg';
+export {default as success} from './status/success.svg';
 
-export {default as circuitBoard} from './circuit-board.svg';
-export {default as satisfaction} from './satisfaction.png';
-export {default as itCrowd} from './IT-Crowd.gif';
+export {default as circuitPattern} from './circuit-pattern.svg';
+export {default as itCrowdGif} from './IT-Crowd.gif';
+
+export {default as card} from './homepage/card.png';
+export {default as wifi} from './homepage/wifi.png';
+export {default as cloud} from './homepage/cloud.png';
+export {default as email} from './homepage/email.png';
+export {default as globe} from './homepage/globe.png';
+export {default as money} from './homepage/money.png';
+export {default as phone} from './homepage/phone.png';
+export {default as rocket} from './homepage/rocket.png';
+export {default as server} from './homepage/server.png';
+export {default as unknown} from './homepage/unknown.png';
+export {default as knife} from './homepage/swiss-knife.png';
+export {default as question} from './homepage/question.png';
+export {default as identity} from './homepage/identity.png';
+export {default as satisfaction} from './homepage/satisfaction.png';